requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
rasterio==1.3.8
numpy==1.24.3
pydantic==2.5.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
httpx==0.25.2
python-dotenv==1.0.0
onnxruntime
google-cloud-storage==2.14.0
